alteração no tipo de retorno dos endpoints

+++ backend/src/main/java/br/gov/sp/fatec/backend/controllers/MessageController.java
@@ -6,15 +6,20 @@ import br.gov.sp.fatec.backend.models.Message;
 import br.gov.sp.fatec.backend.repositories.ConversationRepository;
 import br.gov.sp.fatec.backend.repositories.MemberRepository;
 import br.gov.sp.fatec.backend.repositories.MessageRepository;
+import br.gov.sp.fatec.backend.utils.GrulyApiExceptionResponse;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -38,6 +43,14 @@ public class MessageController {
   @Autowired
   private MemberRepository memberRepository;
 
+  @ExceptionHandler(MissingServletRequestParameterException.class)
+  public ResponseEntity<GrulyApiExceptionResponse<Message>> handleMissingParams(MissingServletRequestParameterException ex) {
+    GrulyApiExceptionResponse<Message> response = new GrulyApiExceptionResponse<Message>();
+    response.addErrorMessage("falta o parâmetro " + ex.getParameterName());
+
+    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+  }
+
   @GetMapping
   @ApiOperation(value = "Retorna uma lista com os dados de todas as mensagens")
   public List<Message> getAllMessages() {
@@ -46,41 +59,89 @@ public class MessageController {
 
   @GetMapping("/{messageId}")
   @ApiOperation(value = "Retorna os dados de uma mensagem")
-  public ResponseEntity<Message> getMessageById(@PathVariable("messageId") long messageId) {
-    return ResponseEntity.ok(messageRepository.findMessageById(messageId));
+  public ResponseEntity<GrulyApiExceptionResponse<Message>> getMessageById(@PathVariable("messageId") long messageId) {
+    GrulyApiExceptionResponse<Message> response = new GrulyApiExceptionResponse<Message>();
+
+    Message fetchedMessage = messageRepository.findMessageById(messageId);
+
+    if(fetchedMessage == null) {
+      response.addErrorMessage("mensagem não encontrada");
+      
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+    }
+
+    response.setData(fetchedMessage);
+
+    return ResponseEntity.ok(response);
   }
 
   @PostMapping
   @ApiOperation(value = "Insere os dados de uma mensagem")
-  public ResponseEntity<Message> insert(@RequestBody Message message, @RequestParam long senderId, @RequestParam long conversationId) {
+  public ResponseEntity<GrulyApiExceptionResponse<Message>> insert(@RequestBody Message message,
+                                                                   @RequestParam("senderId") long senderId,
+                                                                   @RequestParam("conversationId") long conversationId) {
+    GrulyApiExceptionResponse<Message> response = new GrulyApiExceptionResponse<Message>();
+
     Member sender = memberRepository.findMemberById(senderId);
-    Conversation conversation = conversationRepository.findConversationById(conversationId);
+    Conversation chat = conversationRepository.findConversationById(conversationId);
+
+    if(sender == null || chat == null) {
+      if(sender == null) response.addErrorMessage("membro não encontrado");
+      if(chat == null) response.addErrorMessage("conversa não encontrada");
+
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+    }
     
     message.setSender(sender);
-    message.setConversation(conversation);
+    message.setConversation(chat);
 
-    return ResponseEntity.ok(messageRepository.save(message));
+    Message newMessage = messageRepository.save(message);
+
+    if(newMessage == null) {
+      response.addErrorMessage("erro ao criar mensagem");
+
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
+    }
+
+    return ResponseEntity.status(HttpStatus.CREATED).build();
   }
 
   @PutMapping("/{messageId}")
   @ApiOperation(value = "Atualiza os dados de uma mensagem")
-  public ResponseEntity<Message> update(@PathVariable("messageId") long messageId,
-                                        @RequestBody Message updatedMessage) {
-    Member sender = memberRepository.findMemberById(updatedMessage.getSender().getId());
-    Conversation conversation = conversationRepository.findConversationById(updatedMessage.getConversation().getId());
+  public ResponseEntity<GrulyApiExceptionResponse<Message>> update(@PathVariable("messageId") long messageId,
+                                                                   @RequestBody Message messageDataToUpdate) {
+    GrulyApiExceptionResponse<Message> response = new GrulyApiExceptionResponse<Message>();
+
     Message message = messageRepository.findMessageById(messageId);
 
-    message.setText(updatedMessage.getText());
-    message.setTimestamp(updatedMessage.getTimestamp());
-    message.setSender(sender);
-    message.setConversation(conversation);
+    if(messageDataToUpdate.getText() != null) message.setText(messageDataToUpdate.getText());
 
-    return ResponseEntity.ok(messageRepository.save(message));
+    Message updatedMessage = messageRepository.save(message);
+
+    if(updatedMessage == null) {
+      response.addErrorMessage("erro ao atualizar os dados da mensagem");
+
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
+    }
+
+    return ResponseEntity.ok().build();
   }
 
   @DeleteMapping("/{messageId}")
   @ApiOperation(value = "Deleta os dados de uma mensagem")
-  public void delete(@PathVariable("messageId") long messageId) {
+  public ResponseEntity<GrulyApiExceptionResponse<Message>> delete(@PathVariable("messageId") long messageId) {
+    GrulyApiExceptionResponse<Message> response = new GrulyApiExceptionResponse<Message>();
+
+    Message messageToDelete = messageRepository.findMessageById(messageId);
+
+    if(messageToDelete == null) {
+      response.addErrorMessage("mensagem não encontrada");
+
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+    }
+
     messageRepository.deleteById(messageId);
+
+    return ResponseEntity.ok().build();
   }
 }
